requirements.txt
pyyaml
pydantic
jinja2
rich
case-converter
python-can
nunavut